refactor: lzssv2 -> lzss
