+++ std/compress/lzss/average_block.hex
