+++ std/compress/lzss/dict_fun
