+++ std/compress/lzss/dict_naive
