+++ std/compress/lzss/suffixarray/sais.go
