+++ std/compress/lzss/suffixarray/sais2.go
