+++ std/compress/lzss/suffixarray/suffixarray.go
